requirements.txt
discord.py
langchain
langchain-community
langchain-google-genai
google-generativeai
python-dotenv
